+++ lib/kitchen/driver/terraform.rb
@@ -205,6 +205,26 @@ class ::Kitchen::Driver::Terraform < ::Kitchen::Driver::Base
 
   include ::Kitchen::Terraform::Configurable
 
+  # This method queries for the names of the action methods which must be run in serial via a shared mutex.
+  #
+  # If the version satisfies the requirement of ~> 3.3 then no names are returned.
+  #
+  # If the version satisfies the requirement of >= 4 then +:create+, +:converge+, +:setup+, and +:destroy+ are returned.
+  #
+  # @param version [::Kitchen::Terraform::Version] the version
+  # @return [::Array<Symbol>] the action method names
+  def self.serial_actions(version: ::Kitchen::Terraform::Version.new)
+    version
+      .if_satisfies requirement: ::Gem::Requirement.new("~> 3.3") do
+        return []
+      end
+
+    version
+      .if_satisfies requirement: ::Gem::Requirement.new(">= 4") do
+        return super()
+      end
+  end
+
   # Applies changes to the state by selecting the test workspace, updating the dependency modules, validating the root
   # module, applying the state changes, and retrieving the state output.
   #